+++ dramatiq/middleware/__init__.py
@@ -1,7 +1,32 @@
-from .age_limit import AgeLimit  # noqa
-from .callbacks import Callbacks  # noqa
-from .middleware import Middleware, MiddlewareError, SkipMessage  # noqa
-from .pipelines import Pipelines  # noqa
-from .prometheus import Prometheus  # noqa
-from .retries import Retries  # noqa
-from .time_limit import TimeLimit, TimeLimitExceeded  # noqa
+# This file is a part of Dramatiq.
+#
+# Copyright (C) 2017,2018 CLEARTYPE SRL <bogdan@cleartype.io>
+#
+# Dramatiq is free software; you can redistribute it and/or modify it
+# under the terms of the GNU Lesser General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or (at
+# your option) any later version.
+#
+# Dramatiq is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+# License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from .age_limit import AgeLimit
+from .callbacks import Callbacks
+from .middleware import Middleware, MiddlewareError, SkipMessage
+from .pipelines import Pipelines
+from .prometheus import Prometheus
+from .retries import Retries
+from .time_limit import TimeLimit, TimeLimitExceeded
+
+__all__ = [
+    # Basics
+    "Middleware", "MiddlewareError", "SkipMessage",
+
+    # Middlewares
+    "AgeLimit", "Callbacks", "Pipelines", "Prometheus", "Retries", "TimeLimit", "TimeLimitExceeded",
+]