+++ dramatiq/rate_limits/__init__.py
@@ -1,6 +1,28 @@
-from .backend import RateLimiterBackend  # noqa
-from .rate_limiter import RateLimiter, RateLimitExceeded  # noqa
+# This file is a part of Dramatiq.
+#
+# Copyright (C) 2017,2018 CLEARTYPE SRL <bogdan@cleartype.io>
+#
+# Dramatiq is free software; you can redistribute it and/or modify it
+# under the terms of the GNU Lesser General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or (at
+# your option) any later version.
+#
+# Dramatiq is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+# License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .bucket import BucketRateLimiter  # noqa
-from .concurrent import ConcurrentRateLimiter  # noqa
-from .window import WindowRateLimiter  # noqa
+from .backend import RateLimiterBackend
+from .rate_limiter import RateLimiter, RateLimitExceeded
+
+from .bucket import BucketRateLimiter
+from .concurrent import ConcurrentRateLimiter
+from .window import WindowRateLimiter
+
+__all__ = [
+    "RateLimiterBackend", "RateLimiter", "RateLimitExceeded",
+    "BucketRateLimiter", "ConcurrentRateLimiter", "WindowRateLimiter",
+]